vkd3d: Fix integer truncation in find_blob_chunk_masked()
aligned_chunk_size was declared as uint32_t but assigned the result
of align(), which operates in size_t arithmetic. On 64-bit systems,
when chunk->size is near UINT32_MAX, the alignment result exceeds
32 bits and is silently truncated to zero, causing the bounds check
to be bypassed and the size counter to never decrease.

Fix by declaring aligned_chunk_size as size_t to match the type of
the size parameter and the return type of align().

+++ libs/vkd3d/cache.c
@@ -304,7 +304,7 @@ static bool vkd3d_serialized_pipeline_stream_entry_validate(const uint8_t *data,
 static const struct vkd3d_pipeline_blob_chunk *find_blob_chunk_masked(const struct vkd3d_pipeline_blob_chunk *chunk,
         size_t size, uint32_t type, uint32_t mask)
 {
-    uint32_t aligned_chunk_size;
+    size_t aligned_chunk_size;
 
     while (size >= sizeof(struct vkd3d_pipeline_blob_chunk))
     {
